moved types folder into src
